Refactor: Controller에서 세션을 체크하는 로직을 책임에 따라 두 개의 메소드로 분할
1. 세션정보를 바탕으로 로그인 상태인지 체크하는 로직을 하나의 메소드로 분할
2. 세션정보를 바탕으로 해당 기능을 실행할 수 있는 유저인지 체크하는 로직을 다른 하나의 메소드로 분할

+++ src/main/java/com/codessquad/qna/controller/AnswerController.java
@@ -25,9 +25,7 @@ public class AnswerController {
 
     @PostMapping
     public String create(@PathVariable Long questionId, String contents, HttpSession session) {
-        if (!HttpSessionUtils.isLoginUser(session)) {
-            return "users/login";
-        }
+        checkSessionUser(session);
 
         answerService.create(questionId, contents, session);
         return "redirect:/questions/" + questionId;
@@ -35,18 +33,23 @@ public class AnswerController {
 
     @DeleteMapping("/{answerId}")
     public String deleteAnswer(@PathVariable long answerId, HttpSession session, Model model) {
+        checkSessionUser(session);
+
         Answer answer = answerService.getOneById(answerId);
-        checkSession(session, answer);
+
+        checkAccessibleSessionUser(session, answer);
 
         answerService.remove(answer);
         return "redirect:/questions/{questionId}";
     }
 
-    private void checkSession(HttpSession session, Answer answer) {
+    private void checkSessionUser(HttpSession session) {
         if (!HttpSessionUtils.isLoginUser(session)) {
             throw new NotLoggedInException();
         }
+    }
 
+    private void checkAccessibleSessionUser(HttpSession session, Answer answer) {
         User sessionUser = HttpSessionUtils.getUserFromSession(session);
         if (!answer.isEqualWriter(sessionUser)) {
             throw new NotLoggedInException("자신의 글만 수정 및 삭제가 가능합니다.");
